moved doc entity to its own component and exporting from docs aspect to stay backward compatible
