+++ components/entities/doc/doc-prop.ts
