+++ components/entities/doc/doc.docs.mdx
@@ -1,0 +1,4 @@
+---
+labels: ['class']
+description: 'A doc class component to handle basic docs related data and methods'
+---
