+++ scopes/docs/docs/default-doc-reader.ts
@@ -4,7 +4,7 @@ import type { Component } from '@teambit/component';
 import type { CompilerMain } from '@teambit/compiler';
 import type { Workspace } from '@teambit/workspace';
 import { DocReader } from './doc-reader';
-import { Doc } from './doc';
+import { Doc } from '@teambit/docs.entities.doc';
 
 export class DefaultDocReader implements DocReader {
   constructor(private pkg: PkgMain, private compiler: CompilerMain, private workspace: Workspace) {}
