+++ scopes/docs/docs/doc-reader.ts
@@ -1,5 +1,5 @@
 import type { Component } from '@teambit/component';
-import { Doc } from './doc';
+import { Doc } from '@teambit/docs.entities.doc';
 
 export interface DocReader {
   /**
