+++ scopes/docs/docs/index.ts
@@ -4,7 +4,12 @@ export type { DocsMain } from './docs.main.runtime';
 export type { DocsUI } from './docs.ui.runtime';
 export type { DocsPreview } from './docs.preview.runtime';
 export type { DocReader } from './doc-reader';
-export { Doc, DocProp, DocPropList } from './doc';
+
+/**
+ * Doc entity.
+ * @deprecated - please dont use this. use directly from @teambit/docs.entities.doc.
+ */
+export { Doc, DocProp, DocPropList } from '@teambit/docs.entities.doc';
 
 export type { TitleBadgeSlot } from './overview';
